fit all entity - int -> Integer

+++ src/main/java/com/tomcat/service/IRoleService.java
@@ -3,5 +3,5 @@ package com.tomcat.service;
 import com.tomcat.dto.RoleDTO;
 
 public interface IRoleService{
-	RoleDTO getRoleDTO(Long id);
+	RoleDTO getRoleDTO(Integer id);
 }
